Feat : 카테고리 캐시 적용
Refactor : 주문 가능한 매장 목록 조회 리팩터링
