+++ src/main/java/com/flab/delivery/service/CategoryService.java
@@ -1,26 +1,24 @@
 package com.flab.delivery.service;
 
 import com.flab.delivery.dto.category.CategoryDto;
-import com.flab.delivery.dto.store.StoreDto;
 import com.flab.delivery.mapper.CategoryMapper;
-import com.flab.delivery.mapper.StoreMapper;
 import lombok.RequiredArgsConstructor;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+
+import static com.flab.delivery.utils.CacheConstants.CATEGORY;
 
 @Service
 @RequiredArgsConstructor
 public class CategoryService {
 
     private final CategoryMapper categoryMapper;
-    private final StoreMapper storeMapper;
 
+    @Cacheable(value = CATEGORY)
     public List<CategoryDto> getCategories() {
         return categoryMapper.findAllCategory();
     }
 
-    public List<StoreDto> getStoreListBy(Long categoryId, Long addressId) {
-        return storeMapper.findStoreListBy(categoryId, addressId);
-    }
 }
